The state root moves to the shell, and resolve walks the rows once
The simplify round. The state root is environment policy, and every other lookup of that kind lives in the shell beside theme, config and glyphs, resolved once and handed down as data; vigia-core stays a function of git state and explicit arguments. Moving it also removes a second copy of the HOME then USERPROFILE rule, since the shell's home_file already carries it. resolve walked the rows up to three times for one decision and now walks them once. The listing checks an entry's name before building its path. Listing loses the derives nothing used. The test's own temp directory guard was a copy of the scratch helper's naming, and both now share one function in the support module.
